+++ src/ncc/ThirdParty/defuse/php-encryption/Crypto.php
@@ -10,8 +10,11 @@ class Crypto
      * Encrypts a string with a Key.
      *
      * @param string $plaintext
-     * @param Key $key
-     * @param bool $raw_binary
+     * @param Key    $key
+     * @param bool   $raw_binary
+     *
+     * @throws Ex\EnvironmentIsBrokenException
+     * @throws \TypeError
      *
      * @return string
      */
@@ -45,11 +48,19 @@ class Crypto
      *
      * @param string $plaintext
      * @param string $password
-     * @param bool $raw_binary
+     * @param bool   $raw_binary
+     *
+     * @throws Ex\EnvironmentIsBrokenException
+     * @throws \TypeError
      *
      * @return string
      */
-    public static function encryptWithPassword($plaintext, $password, $raw_binary = false)
+    public static function encryptWithPassword(
+        $plaintext,
+        #[\SensitiveParameter]
+        $password,
+        $raw_binary = false
+    )
     {
         if (!\is_string($plaintext)) {
             throw new \TypeError(
@@ -124,7 +135,12 @@ class Crypto
      *
      * @return string
      */
-    public static function decryptWithPassword($ciphertext, $password, $raw_binary = false)
+    public static function decryptWithPassword(
+        $ciphertext,
+        #[\SensitiveParameter]
+        $password,
+        $raw_binary = false
+    )
     {
         if (!\is_string($ciphertext)) {
             throw new \TypeError(
@@ -160,7 +176,11 @@ class Crypto
      *
      * @return string
      */
-    public static function legacyDecrypt($ciphertext, $key)
+    public static function legacyDecrypt(
+        $ciphertext,
+        #[\SensitiveParameter]
+        $key
+    )
     {
         if (!\is_string($ciphertext)) {
             throw new \TypeError(
@@ -372,7 +392,13 @@ class Crypto
      *
      * @return string
      */
-    protected static function plainEncrypt($plaintext, $key, $iv)
+    protected static function plainEncrypt(
+        $plaintext,
+        #[\SensitiveParameter]
+        $key,
+        #[\SensitiveParameter]
+        $iv
+    )
     {
         Core::ensureConstantExists('OPENSSL_RAW_DATA');
         Core::ensureFunctionExists('openssl_encrypt');
@@ -402,7 +428,14 @@ class Crypto
      *
      * @return string
      */
-    protected static function plainDecrypt($ciphertext, $key, $iv, $cipherMethod)
+    protected static function plainDecrypt(
+        $ciphertext,
+        #[\SensitiveParameter]
+        $key,
+        #[\SensitiveParameter]
+        $iv,
+        $cipherMethod
+    )
     {
         Core::ensureConstantExists('OPENSSL_RAW_DATA');
         Core::ensureFunctionExists('openssl_decrypt');
@@ -431,7 +464,12 @@ class Crypto
      *
      * @return bool
      */
-    protected static function verifyHMAC($expected_hmac, $message, $key)
+    protected static function verifyHMAC(
+        $expected_hmac,
+        $message,
+        #[\SensitiveParameter]
+        $key
+    )
     {
         $message_hmac = \hash_hmac(Core::HASH_FUNCTION_NAME, $message, $key, true);
         return Core::hashEquals($message_hmac, $expected_hmac);